api/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==20.8b1
boto3==1.18.6
botocore==1.21.6
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-polymorphic==3.0.0
django-rest-polymorphic==0.1.9
django-storages==1.11.1
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.8.1
Pillow==8.3.1
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
regex==2021.4.4
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.6
django-filter==2.4.0
